Remove unused ‘static’ applicator
Note: If you want to apply static methods to Schema, then for now simply apply them at a code level:

    var Hero = schema.new(‘Hero’).prop(‘power’);
	Hero.staticMethod = function(){ console.log(‘derp’); }

If we get to the point where we need to store these statics for bootstrapping into an app, we’ll deal with that then. Right now it’s added complexity for no gain.